+++ benches/lookup.rs
@@ -34,7 +34,7 @@ where
     for ip in ips.iter() {
         let result = reader.lookup(*ip).unwrap();
         if result.has_data() {
-            let _: geoip2::City = result.decode().unwrap();
+            let _: geoip2::City = result.decode().unwrap().unwrap();
         }
     }
 }
@@ -47,7 +47,7 @@ where
     ips.par_iter().for_each(|ip| {
         let result = reader.lookup(*ip).unwrap();
         if result.has_data() {
-            let _: geoip2::City = result.decode().unwrap();
+            let _: geoip2::City = result.decode().unwrap().unwrap();
         }
     });
 }